+++ packages/codegen/src/EntityDbMetadata.ts
@@ -229,6 +229,7 @@ export class EntityDbMetadata {
   updatedAt: PrimitiveField | undefined;
   deletedAt: PrimitiveField | undefined;
   baseClassName: string | undefined;
+  subTypes: EntityDbMetadata[] = [];
   inheritanceType: "sti" | "cti" | undefined;
   /** This will only be set on the base meta. */
   stiDiscriminatorField?: string;